+++ src/app/features/preprints/mappers/index.ts
@@ -1,1 +1,2 @@
+export * from './preprint-providers.mapper';
 export * from './preprints.mapper';
