+++ src/app/features/preprints/services/index.ts
@@ -1,3 +1,6 @@
-export { ContributorsService } from './contributors.service';
+export { PreprintContributorsService } from './preprint-contributors.service';
+export { PreprintFilesService } from './preprint-files.service';
+export { PreprintLicensesService } from './preprint-licenses.service';
+export { PreprintProvidersService } from './preprint-providers.service';
 export { PreprintsService } from './preprints.service';
 export { PreprintsFiltersOptionsService } from './preprints-resource-filters.service';
