+++ src/app/features/preprints/services/preprint-licenses.service.ts
@@ -18,7 +18,7 @@ import { environment } from 'src/environments/environment';
 @Injectable({
   providedIn: 'root',
 })
-export class LicensesService {
+export class PreprintLicensesService {
   private apiUrl = environment.apiUrl;
   private readonly jsonApiService = inject(JsonApiService);
